nuts

+++ src/routes/+page.server.ts
@@ -2,15 +2,15 @@ import { fail, type Actions } from "@sveltejs/kit";
 
 import { prisma } from "../lib/server/prisma";
 import type { PageServerLoad } from "./$types";
-import { mintExists, validateMintUrl } from "../util/util";
+import { claimToken, mintExists, validateMintUrl } from "../util/util";
 
 export const load: PageServerLoad = async () => {
-    return { mints: await prisma.mint.findMany() }
+    return { mints: await prisma.mint.findMany({ include: { uptime: true } }) }
 };
 
 export const actions: Actions = {
     createMint: async ({ request }) => {
-        const { url } = Object.fromEntries(await request.formData()) as { url: string }
+        const { url, token } = await request.json()
 
         if (!validateMintUrl(url)) {
             return fail(400, { message: 'Mint url not valid' })
@@ -21,11 +21,17 @@ export const actions: Actions = {
         }
 
         if (await mintExists(url)) {
-            return fail(400, { message: 'Mint already' })
+            return fail(400, { message: 'Mint already exists' })
         }
 
+        if (!(await claimToken(token, 21))) {
+            return fail(400, { message: 'provided invalid or not enough ecash' })
+        }
+        let id
         try {
-            await prisma.mint.create({ data: { url } })
+            const mint = await prisma.mint.create({ data: { url } })
+            console.log(mint)
+            id = mint.id
         }
         catch (err) {
             console.error(err)
@@ -33,21 +39,28 @@ export const actions: Actions = {
         }
 
         return {
-            status: 201
+            status: 201,
+            id
         }
     },
 
-    deleteMint: async ({ url }) => {
-        const id = url.searchParams.get('id')
+    deleteMint: async ({ request }) => {
+        const { id, token } = await request.json()
+
         if (!id) {
             return fail(400, { message: 'no mint id specified' })
         }
-
+        if (!(await claimToken(token, 210))) {
+            return fail(400, { message: 'provided invalid or not enough ecash' })
+        }
         try {
             await prisma.mint.delete({ where: { id: Number(id) } })
         } catch (err) {
             console.error(err)
             return fail(500, { message: 'could not delete mint item' })
+        }
+        return {
+            status: 200,
         }
     }
 };
@@ -65,3 +78,4 @@ const canLoadKeys = async (url: string): Promise<boolean> => {
         return false
     }
 }
+
